typo fixed

+++ parse_vector.c
@@ -5,7 +5,7 @@
 
 #define SIZE INT_MAX // INT_MAX = +2147483647
 
-int16_t *parse_vector(int32_t n) {
+int16_t *parse_vector(int64_t n) {
 	int16_t *vector = malloc(SIZE * sizeof(int16_t));
 	int16_t length = (int) floor(log10(n)); // log_base10(integer) -> integer lenght
 	for (int i = 0; i <= length; i++) {
